requirements.txt
# Core (install torch/torchvision separately for your CUDA)
ultralytics>=8.3.0
roboflow>=1.1.0

transformers>=4.42.0
accelerate>=0.30.0
datasets>=2.20.0

pyyaml>=6.0.1
numpy>=1.24
pandas>=2.0
tqdm>=4.66
pillow>=10.0
opencv-python>=4.8
matplotlib>=3.8
tensorboard>=2.16
pycocotools>=2.0.7

timm>=1.0.22
scipy
tabulate

# Optional (only needed for GLIP; see README)
# detectron2 (install separately)
